+++ src/runlength_late.lua
@@ -1,6 +1,6 @@
 --#region Run result subtitle
 
-modutil.mod.Path.Context.Wrap.Static("ShowRunHistory", function ( screen, button )
+modutil.mod.Path.Context.Env("ShowRunHistory", function ( screen, button )
     modutil.mod.Path.Wrap("ModifyTextBox", function (base, args)
         args = args or {}
         if args.Text == "RunHistoryScreen_DreamBiomeVisitOrder" and type(args.LuaValue) == "table" and #args.LuaValue > 4 then
@@ -10,7 +10,7 @@ modutil.mod.Path.Context.Wrap.Static("ShowRunHistory", function ( screen, button
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("OpenRunClearScreen", function (  )
+modutil.mod.Path.Context.Env("OpenRunClearScreen", function (  )
     modutil.mod.Path.Wrap("RunClearMessagePresentation", function (base, screen, message, tooltipData)
         if message == "ClearDreamRun" and type(tooltipData) == "table" and #tooltipData > 4 then
             message = message .. #tooltipData
@@ -62,7 +62,7 @@ local NPC_fucntions = {
 }
 
 for _, functionName in ipairs(NPC_fucntions) do
-    modutil.mod.Path.Context.Wrap.Static(functionName, function ()
+    modutil.mod.Path.Context.Env(functionName, function ()
         modutil.mod.Path.Wrap("OpenUpgradeChoiceMenu", function (base, source, args)
             return mod.OpenUpgradeChoiceMenu_NPC(base, source, args)
         end)