+++ src/visage_late.lua
@@ -1,4 +1,4 @@
-modutil.mod.Path.Context.Wrap.Static("HecateKillPresentation", function ( unit, args )
+modutil.mod.Path.Context.Env("HecateKillPresentation", function ( unit, args )
     modutil.mod.Path.Wrap("AddOutline", function (base, args)
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model then
             return
@@ -14,7 +14,7 @@ modutil.mod.Path.Context.Wrap.Static("HecateKillPresentation", function ( unit, 
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("HecateBattleStart", function ( hecate, args )
+modutil.mod.Path.Context.Env("HecateBattleStart", function ( hecate, args )
     modutil.mod.Path.Wrap("AddOutline", function (base, args)
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model then
             return
@@ -32,7 +32,7 @@ modutil.mod.Path.Context.Wrap.Static("HecateBattleStart", function ( hecate, arg
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("InfestedCerberusHorribleRaceConditionForTempPresentation", function ( unit )
+modutil.mod.Path.Context.Env("InfestedCerberusHorribleRaceConditionForTempPresentation", function ( unit )
      modutil.mod.Path.Wrap("AddOutline", function (base, args)
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model then
             return
@@ -48,7 +48,7 @@ modutil.mod.Path.Context.Wrap.Static("InfestedCerberusHorribleRaceConditionForTe
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("CerberusStageEnter", function ( unit )
+modutil.mod.Path.Context.Env("CerberusStageEnter", function ( unit )
     modutil.mod.Path.Wrap("SetThingProperty", function (base, args)
         args = args or {}
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model and args.Property == "GrannyTexture" and args.Value == "GR2/InfestedCerberusDreamEM2_Color" then
@@ -58,7 +58,7 @@ modutil.mod.Path.Context.Wrap.Static("CerberusStageEnter", function ( unit )
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("TyphonHeadSummonPresentation", function ( unit )
+modutil.mod.Path.Context.Env("TyphonHeadSummonPresentation", function ( unit )
     modutil.mod.Path.Wrap("SetThingProperty", function (base, args)
         args = args or {}
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model and args.Property == "GrannyTexture" and string.match(args.Value, "_Color") then
@@ -68,11 +68,21 @@ modutil.mod.Path.Context.Wrap.Static("TyphonHeadSummonPresentation", function ( 
     end)
 end)
 
-modutil.mod.Path.Context.Wrap.Static("TyphonHeadStageTransition", function ( unit )
+modutil.mod.Path.Context.Env("TyphonHeadStageTransition", function ( unit )
     modutil.mod.Path.Wrap("SetThingProperty", function (base, args)
         args = args or {}
         if game.CurrentRun.IsDreamRun and config.disable_visage_forms.model and args.Property == "GrannyTexture" and string.match(args.Value, "_Color") then
             args.Value = string.gsub(args.Value, "Dream", "")
+        end
+        return base(args)
+    end)
+end)
+
+modutil.mod.Path.Context.Env("FakeDeathTyphonEntrance", function ( unit )
+    modutil.mod.Path.Wrap("SetThingProperty", function (base, args)
+        args = args or {}
+        if args.Value == "GR2/TyphonHeadDreamStage3_Color" and game.CurrentRun.IsDreamRun and config.disable_visage_forms.model and args.Property == "GrannyTexture" then
+            args.Value = "GR2/TyphonHeadStage3_Color"
         end
         return base(args)
     end)